+++ .travis.yml
@@ -10,6 +10,7 @@ rvm:
   - 2.2.6
   - 2.3.3
   - 2.4.0
+  - 2.5.1
   - jruby-19mode
   - jruby-9.1.7.0
 env:
