+++ README.md
@@ -22,27 +22,22 @@ Or install it yourself as:
 
 ## Usage
 
-Initialise a `Fog::Storage` object using local provider
+Initialise a `Fog::Local::Storage` object:
 
 ```ruby
-
-storage = Fog::Storage.new({
-  :local_root => '~/fog',
-  :provider   => 'Local'
-})
-
+storage = Fog::Local::Storage.new(local_root: '~/fog')
 ```
 
 This can then be used like any other [Fog storage](http://fog.io/storage/).
 
 ```ruby
-directory = storage.directories.create(:key => 'data')
-directory.files.create(:body => 'Hello World!', :key  => 'hello_world.txt')
+directory = storage.directories.create(key: 'data')
+directory.files.create(body: 'Hello World!', key: 'hello_world.txt')
 ```
 
 ## Contributing
 
-1. Fork it ( https://github.com/fog/fog-local/fork )
+1. Fork it ( https://github.com/fog/fog-local/fork)
 2. Create your feature branch (`git checkout -b my-new-feature`)
 3. Commit your changes (`git commit -am 'Add some feature'`)
 4. Push to the branch (`git push origin my-new-feature`)
